requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
jinja2>=3.1.2
httpx>=0.25.0
supabase>=2.3.0
python-dateutil>=2.8.2
mangum>=0.17.0
